Function slow_primes::estimate_nth_prime

pub fn estimate_nth_prime(n: u64) -> (u64, u64)

Gives estimated bounds for pn, the nth prime number, 1-indexed (i.e. p1 = 2, p2 = 3).

That is, if (a,b) = estimate_nth_prime(n), a ≤ pn ≤ b. The bounds used are proved in [1] and [2, Théorèmes 1.6–1.8], and are summarised in [2, pp. 14–15].

[1]: Massias, Jean-Pierre; Robin, Guy. "Bornes effectives pour certaines fonctions concernant les nombres premiers." Journal de théorie des nombres de Bordeaux 8.1 (1996): 215-242.

[2]: Dusart, Pierre. "Autour de la fonction qui compte le nombre de nombres premiers." PhD diss., Université de Limoges, 1998.